lib/job.rb: modify the filter call condition to check top level values only
When the parameter with the same name as filter is set in other
programs, the related filter will be called but the value read is nil.
E.g:
```
suite: memtier
testcase: memtier
category: benchmark

cpu_node_bind: even

nr_processes: 4

if role client:
  memtier:
    nr_threads: 1
```
In the above example, nr_threads is defined under memtier, so
filte/nr_threads will be called, but nr_threads will become
memtier_nr_threads in the incoming value, which will result in the value
of $nr_threads obtained in filte/nr_threads is nil.
Therefore, the call condition of the filter is modified. Only when the
parameter is defined at the top level, the relevant filter will be
called.

Signed-off-by: Zhou Hao <haox.a.zhou@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>